ftx_feed/requirements.txt
grpcio==1.48.1
websockets==10.3
grpcio-tools==1.48.1